feat: DraftPopupArea
